Made corrections after review; merged into npo.py
